Rough code for Sentiment Scoring w. GPT
Assuming you have a folder with video frames, this will compare each frame to the next and produce an analysis dictionary that you can use for front-end pop-up behavior. 
I'm currently setting up threading so that we can read in frames while we simultaneously process them. The looping behavior & capture logic will be changing but the dictionary structure is likely going to remain the same. You can use that dictionary structure to build popup behavior & get a feel for the video-analysis abilities.