Rough code for Speech Rate Analysis w Google
This code contains a minimally viable threaded implementation for audio-processing and speech rate analysis. There are some errors in the thread overlap since some text seems to get lost, and the time controls which dictate the regular interval for chunking do not seem properly enforced. Additionally, it is wrapped in an interface which is unnecessary for integration with the larger system. I intend to revisit this code today but I wanted to make it available for anyone who wants to get their hands on it.