Rough code for PDF Analysis w PyMuPDF
Uploading a pdf will perform a page-by-page analysis using PyMuPDF for speed and implementation simplicity. Current output offers: wordcount and sentiment analysis. Sentiment analysis is performed with the same library as the one we are using for audio-evaluation - please do not change this. The values output for sentiment are the confidence levels for general tone of the language. The variations are: negative, neutral, positive, and compound. Compound represents the normalized sum of the other three scores.  

Interface included for testing purposes. Safe to remove.